Should work better

+++ packages/web/lib/reg-task/taskqueue.class.php
@@ -376,10 +376,12 @@ class TaskQueue extends TaskingElement
         self::$FOGSSH->username = $this->StorageNode->get('user');
         self::$FOGSSH->password = $this->StorageNode->get('pass');
         self::$FOGSSH->host = $this->StorageNode->get('ip');
-        self::$FOGSSH->connect();
-        self::$FOGSSH->delete($dest);
-        self::$FOGSSH->sftp_rename($src, $dest);
-        self::$FOGSSH->disconnect();
+        if (self::$FOGSSH->connect()) {
+            sleep(2);
+            self::$FOGSSH->delete($dest);
+            self::$FOGSSH->sftp_rename($src, $dest);
+            self::$FOGSSH->disconnect();
+        }
         if ($this->Image->get('format') == 1) {
             $this->Image
                 ->set('format', 0)
